+++ packages/cli/src/base.ts
@@ -20,7 +20,6 @@ export abstract class BaseCommand extends Command {
 
   private _web3: Web3 | null = null
   private _kit: ContractKit | null = null
-  private _metadataManager: MetadataManager | null = null
 
   // This is required since we wrap the provider with a debug provider and
   // there is no way to unwrap the provider afterwards.
@@ -41,13 +40,6 @@ export abstract class BaseCommand extends Command {
       this._kit = newKitFromWeb3(this.web3)
     }
     return this._kit
-  }
-
-  get metadataManager() {
-    if (!this._metadataManager) {
-      this._metadataManager = new MetadataManager()
-    }
-    return this._metadataManager
   }
 
   async init() {
