+++ packages/contractkit/src/identity/index.ts
@@ -1,2 +1,1 @@
-export * from './metadata-manager'
-export * from './types'
+export * from './metadata'
